M4/role-playing-chat/requirements.txt
google-genai
python-dotenv
colorama
